// src/app/models/property.ts
import { PropertyI } from "./property-i";

export class Property implements PropertyI {
  Id: number;
  SellRent: number;
  Name: string;
  PType: string;
  BHK: number;
  FType: string;
  Price: number;
  BuiltArea: number;
  CarpetArea: number;
  Image: string;
  Address: string;
  Address2: string;
  Address3: string;
  City: string;
  Description: string;
  FloorNo: string;
  TotalFloor: string;
  RTM: number;
  AOP: number;
  Bathrooms: number;
  MainEntrance: string;
  Gated: number;
  Security: number;
  Maintenance: number;
  Possession: string;
  PostedOn: string
}